+++ pkg/coremodel/dashboard/x/looseygoosey/dashboard.go
@@ -1,4 +1,4 @@
-package looseygoosey
+package schemaless
 
 import (
 	"io"
