+++ pkg/coremodel/dashboard/x/looseygoosey/dashboard_test.go
@@ -1,4 +1,4 @@
-package looseygoosey
+package schemaless
 
 import (
 	"encoding/json"
