+++ pkg/coremodel/dashboard/x/looseygoosey/ds_lookup.go
@@ -1,4 +1,4 @@
-package looseygoosey
+package schemaless
 
 import (
 	"context"
