+++ pkg/coremodel/dashboard/x/looseygoosey/reference.go
@@ -1,4 +1,4 @@
-package looseygoosey
+package schemaless
 
 import (
 	"fmt"
