+++ pkg/coremodel/dashboard/x/looseygoosey/summary.go
@@ -1,4 +1,4 @@
-package looseygoosey
+package schemaless
 
 import (
 	"bytes"
