+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/all-selected-multi-datasource-variable.json
