+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/all-selected-single-datasource-variable-info.json
