+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/dash_labels.jsonc
