+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/dash_raw.jsonc
