+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/dash_references.jsonc
