+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/dash_summary.jsonc
