+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/datasource-variable-info.json
