+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/datasource-variable-no-curly-braces-info.json
