+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/datasource-variable.json
