+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/default-datasource-variable-info.json
