+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/devdash-all-panels-info.json
