+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/devdash-graph-shared-tooltips-info.json
