+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/empty-datasource-variable-info.json
