+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/gdev-walk-graph-gradient-area-fills.json
@@ -1,0 +1,93 @@
+{
+  "name": "Panel Tests - Graph - Gradient Area Fills",
+  "labels": {
+    "gdev": "",
+    "graph": "",
+    "panel-tests": ""
+  },
+  "URL": "/d/graph-gradient-area-fills.json/panel-tests-graph-gradient-area-fills",
+  "fields": {
+    "schemaVersion": 18
+  },
+  "nested": [
+    {
+      "uid": "graph-gradient-area-fills.json#2",
+      "kind": "panel",
+      "name": "Req/s",
+      "URL": "/d/graph-gradient-area-fills.json/panel-tests-graph-gradient-area-fills?viewPanel=2",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph-gradient-area-fills.json#11",
+      "kind": "panel",
+      "name": "Req/s",
+      "URL": "/d/graph-gradient-area-fills.json/panel-tests-graph-gradient-area-fills?viewPanel=11",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph-gradient-area-fills.json#7",
+      "kind": "panel",
+      "name": "Memory",
+      "URL": "/d/graph-gradient-area-fills.json/panel-tests-graph-gradient-area-fills?viewPanel=7",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph-gradient-area-fills.json#10",
+      "kind": "panel",
+      "name": "Req/s",
+      "URL": "/d/graph-gradient-area-fills.json/panel-tests-graph-gradient-area-fills?viewPanel=10",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    }
+  ],
+  "references": [
+    {
+      "kind": "ds",
+      "type": "default.type",
+      "UID": "default.uid"
+    },
+    {
+      "kind": "panel",
+      "type": "graph"
+    }
+  ]
+}
