+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/gdev-walk-graph-shared-tooltips.json
@@ -1,0 +1,181 @@
+{
+  "name": "Panel Tests - shared tooltips",
+  "labels": {
+    "gdev": "",
+    "graph-ng": "",
+    "panel-tests": ""
+  },
+  "URL": "/d/graph-shared-tooltips.json/panel-tests-shared-tooltips",
+  "fields": {
+    "schemaVersion": 28
+  },
+  "nested": [
+    {
+      "uid": "graph-shared-tooltips.json#4",
+      "kind": "panel",
+      "name": "two units",
+      "URL": "/d/graph-shared-tooltips.json/panel-tests-shared-tooltips?viewPanel=4",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "timeseries"
+        }
+      ]
+    },
+    {
+      "uid": "graph-shared-tooltips.json#13",
+      "kind": "panel",
+      "name": "Speed vs Temperature (XY)",
+      "URL": "/d/graph-shared-tooltips.json/panel-tests-shared-tooltips?viewPanel=13",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "xychart"
+        },
+        {
+          "kind": "transform",
+          "type": "organize"
+        },
+        {
+          "kind": "transform",
+          "type": "seriesToColumns"
+        }
+      ]
+    },
+    {
+      "uid": "graph-shared-tooltips.json#2",
+      "kind": "panel",
+      "name": "Cursor info",
+      "URL": "/d/graph-shared-tooltips.json/panel-tests-shared-tooltips?viewPanel=2",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "debug"
+        }
+      ]
+    },
+    {
+      "uid": "graph-shared-tooltips.json#5",
+      "kind": "panel",
+      "name": "Only temperature",
+      "URL": "/d/graph-shared-tooltips.json/panel-tests-shared-tooltips?viewPanel=5",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "timeseries"
+        }
+      ]
+    },
+    {
+      "uid": "graph-shared-tooltips.json#9",
+      "kind": "panel",
+      "name": "Only Speed",
+      "URL": "/d/graph-shared-tooltips.json/panel-tests-shared-tooltips?viewPanel=9",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "timeseries"
+        }
+      ]
+    },
+    {
+      "uid": "graph-shared-tooltips.json#11",
+      "kind": "panel",
+      "name": "Panel Title",
+      "URL": "/d/graph-shared-tooltips.json/panel-tests-shared-tooltips?viewPanel=11",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "timeseries"
+        }
+      ]
+    },
+    {
+      "uid": "graph-shared-tooltips.json#8",
+      "kind": "panel",
+      "name": "flot panel (temperature)",
+      "URL": "/d/graph-shared-tooltips.json/panel-tests-shared-tooltips?viewPanel=8",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph-shared-tooltips.json#10",
+      "kind": "panel",
+      "name": "flot panel (no units)",
+      "URL": "/d/graph-shared-tooltips.json/panel-tests-shared-tooltips?viewPanel=10",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    }
+  ],
+  "references": [
+    {
+      "kind": "ds",
+      "type": "default.type",
+      "UID": "default.uid"
+    },
+    {
+      "kind": "panel",
+      "type": "debug"
+    },
+    {
+      "kind": "panel",
+      "type": "graph"
+    },
+    {
+      "kind": "panel",
+      "type": "timeseries"
+    },
+    {
+      "kind": "panel",
+      "type": "xychart"
+    }
+  ]
+}
