+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/gdev-walk-graph-time-regions.json
@@ -1,0 +1,110 @@
+{
+  "name": "Panel Tests - Graph Time Regions",
+  "labels": {
+    "gdev": "",
+    "graph": "",
+    "panel-tests": ""
+  },
+  "URL": "/d/graph-time-regions.json/panel-tests-graph-time-regions",
+  "fields": {
+    "schemaVersion": 18
+  },
+  "nested": [
+    {
+      "uid": "graph-time-regions.json#2",
+      "kind": "panel",
+      "name": "Business Hours",
+      "URL": "/d/graph-time-regions.json/panel-tests-graph-time-regions?viewPanel=2",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "testdata",
+          "UID": "PD8C576611E62080A"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph-time-regions.json#4",
+      "kind": "panel",
+      "name": "Sunday's 20-23",
+      "URL": "/d/graph-time-regions.json/panel-tests-graph-time-regions?viewPanel=4",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "testdata",
+          "UID": "PD8C576611E62080A"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph-time-regions.json#3",
+      "kind": "panel",
+      "name": "Each day of week",
+      "URL": "/d/graph-time-regions.json/panel-tests-graph-time-regions?viewPanel=3",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "testdata",
+          "UID": "PD8C576611E62080A"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph-time-regions.json#5",
+      "kind": "panel",
+      "name": "05:00",
+      "URL": "/d/graph-time-regions.json/panel-tests-graph-time-regions?viewPanel=5",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "testdata",
+          "UID": "PD8C576611E62080A"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph-time-regions.json#7",
+      "kind": "panel",
+      "name": "From 22:00 to 00:30 (crossing midnight)",
+      "URL": "/d/graph-time-regions.json/panel-tests-graph-time-regions?viewPanel=7",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "testdata",
+          "UID": "PD8C576611E62080A"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    }
+  ],
+  "references": [
+    {
+      "kind": "ds",
+      "type": "testdata",
+      "UID": "PD8C576611E62080A"
+    },
+    {
+      "kind": "panel",
+      "type": "graph"
+    }
+  ]
+}
