+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/gdev-walk-graph_y_axis.json
@@ -1,0 +1,177 @@
+{
+  "name": "Panel Tests - Graph - Y axis ticks",
+  "labels": {
+    "gdev": "",
+    "panel-tests": ""
+  },
+  "URL": "/d/graph_y_axis.json/panel-tests-graph-y-axis-ticks",
+  "fields": {
+    "schemaVersion": 19
+  },
+  "nested": [
+    {
+      "uid": "graph_y_axis.json#7",
+      "kind": "panel",
+      "name": "Data from 0 - 10K (unit short)",
+      "URL": "/d/graph_y_axis.json/panel-tests-graph-y-axis-ticks?viewPanel=7",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph_y_axis.json#5",
+      "kind": "panel",
+      "name": "Data from 0 - 10K (unit bytes metric)",
+      "URL": "/d/graph_y_axis.json/panel-tests-graph-y-axis-ticks?viewPanel=5",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph_y_axis.json#4",
+      "kind": "panel",
+      "name": "Data from 0 - 10K (unit bytes IEC)",
+      "URL": "/d/graph_y_axis.json/panel-tests-graph-y-axis-ticks?viewPanel=4",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph_y_axis.json#2",
+      "kind": "panel",
+      "name": "Data from 0 - 10K (unit short)",
+      "URL": "/d/graph_y_axis.json/panel-tests-graph-y-axis-ticks?viewPanel=2",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph_y_axis.json#3",
+      "kind": "panel",
+      "name": "Data from 0.0002 - 0.001 (unit short)",
+      "URL": "/d/graph_y_axis.json/panel-tests-graph-y-axis-ticks?viewPanel=3",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph_y_axis.json#6",
+      "kind": "panel",
+      "name": "Data from 12000 - 30000 (unit ms)",
+      "URL": "/d/graph_y_axis.json/panel-tests-graph-y-axis-ticks?viewPanel=6",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph_y_axis.json#9",
+      "kind": "panel",
+      "name": "Data from 0 - 1B (unit short)",
+      "URL": "/d/graph_y_axis.json/panel-tests-graph-y-axis-ticks?viewPanel=9",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph_y_axis.json#10",
+      "kind": "panel",
+      "name": "Data from 0 - 1B (unit bytes)",
+      "URL": "/d/graph_y_axis.json/panel-tests-graph-y-axis-ticks?viewPanel=10",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    },
+    {
+      "uid": "graph_y_axis.json#8",
+      "kind": "panel",
+      "name": "Data from 12000 - 30000 (unit ms)",
+      "URL": "/d/graph_y_axis.json/panel-tests-graph-y-axis-ticks?viewPanel=8",
+      "references": [
+        {
+          "kind": "ds",
+          "type": "default.type",
+          "UID": "default.uid"
+        },
+        {
+          "kind": "panel",
+          "type": "graph"
+        }
+      ]
+    }
+  ],
+  "references": [
+    {
+      "kind": "ds",
+      "type": "default.type",
+      "UID": "default.uid"
+    },
+    {
+      "kind": "panel",
+      "type": "graph"
+    }
+  ]
+}
