+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/mixed-datasource-with-variable-info.json
