+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/repeated-datasource-variables-info.json
