+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/repeated-datasource-variables-with-default.json
