+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/repeated-datasource-variables.json
