+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/special-datasource-types-info.json
