+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/special-datasource-types.json
