+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/string-datasource-variable-info.json
