+++ pkg/coremodel/dashboard/x/looseygoosey/testdata/string-datasource-variable.json
